Fix most npm vulnerabilities
There are a few npm warnings remaining but they rely on other packages to update like React:
https://github.com/facebook/create-react-app/issues/11054

For now, I've switched the build to use Yarn which allows us to use resolutions to include the newer versions of these insecure packages.

The end result is that npm will still say there are fixes required, but the actual packages being used to build this package are the fixed versions.